% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{as_applied_s_rate}
\alias{as_applied_s_rate}
\title{As-applied seed rate}
\format{
sf
A data frame with 1,041 rows and 2 columns:
\describe{
  \item{seed_rate}{Seed rate in 1000}
  \item{geometry}{geometry}
}
}
\source{
privately obtained
}
\usage{
as_applied_s_rate
}
\description{
As-applied seed rate in 1000
}
\keyword{datasets}
