% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co_counties}
\alias{co_counties}
\title{Colorado county boundary data}
\format{
`sf`
An sf with 2 rows and 1 columns:
\describe{
  \item{statefp}{State FIP}
  \item{countyfp}{County FIP}
  \item{name}{County name}
  \item{geometry}{geometry}
}
}
\source{
<https://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2019/TGRSHP2019_TechDoc.pdf>
}
\usage{
co_counties
}
\description{
Colorado county boundary as an sf object
}
\keyword{datasets}
