% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corn_acres_ne}
\alias{corn_acres_ne}
\title{Corn harvested acres in Nebraska}
\format{
`sf`
An sf with 6205 rows and 4 columns:
\describe{
  \item{year}{year}
  \item{county_code}{County FIP}
  \item{acre}{harvested corn acres}
  \item{geometry}{geometry}
}
}
\source{
<https://www.nass.usda.gov/Quick_Stats/>
}
\usage{
corn_acres_ne
}
\description{
Corn harvested acres in Nebraska
}
\keyword{datasets}
