% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corn_yield}
\alias{corn_yield}
\title{Corn yield}
\format{
sf
An sf object with 3830 rows and 2 columns:
\describe{
  \item{geometry}{geometry}
  \item{yield}{corn yield}
}
}
\source{
privately obtained
}
\usage{
corn_yield
}
\description{
Corn yield from yield monitor
}
\keyword{datasets}
