% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fairway_grid}
\alias{fairway_grid}
\title{Fairway plot data}
\format{
`sf`
An sf with 42 rows and 2 columns:
\describe{
  \item{grid}{grid ID}
  \item{geometry}{geometry}
}
}
\source{
privately obtained
}
\usage{
fairway_grid
}
\description{
Fairway plot polygons saved as an sf object
}
\keyword{datasets}
