% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hp_boundary}
\alias{hp_boundary}
\title{High-Plains Aquifer boundary data}
\format{
`sf`
An sf with 2 rows and 1 columns:
\describe{
  \item{geometry}{geometry}
}
}
\source{
<https://www.sciencebase.gov/catalog/item/6314061bd34e36012efa397b>
}
\usage{
hp_boundary
}
\description{
High-Plains Aquifer boundary polygon data saved as an sf object
}
\keyword{datasets}
