% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huc_ia}
\alias{huc_ia}
\title{Hydrologic Units}
\format{
sf
A data frame with 29 rows and 2 columns:
\describe{
  \item{huc_code}{HUC ID}
  \item{geometry}{geometry}
}
}
\source{
<"https://www.sciencebase.gov/catalog/item/631405c4d34e36012efa315f">
}
\usage{
huc_ia
}
\description{
Hydrologic units that covers Iowa
}
\keyword{datasets}
