% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_corn_acres}
\alias{ia_corn_acres}
\title{Corn acreage in Iowa}
\format{
sf
A data frame with 1,000 rows and 3 columns:
\describe{
  \item{county_code}{County FIPS number}
  \item{year}{year}
  \item{acres}{Corn acres}
  \item{geometry}{geometry}
}
}
\source{
USDA NASS
}
\usage{
ia_corn_acres
}
\description{
Corn acreage in Iowa by county in 2018
}
\keyword{datasets}
