% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_nitrogen}
\alias{ia_nitrogen}
\title{Fake nitrogen use by county in Iowa}
\format{
sf
An sf with 99 rows and 3 columns:
\describe{
  \item{countyfp}{County FIPS number}
  \item{geometry}{geometry}
  \item{nitrogen_rate}{Nitrogen rate (lb/acre)}
}
}
\source{
USDA NASS
}
\usage{
ia_nitrogen
}
\description{
Fake nitrogen use by county in Iowa
}
\keyword{datasets}
