% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines}
\alias{lines}
\title{Lines}
\format{
sf
A data frame with 2 rows and 2 columns:
\describe{
  \item{x}{geometry}
  \item{line_name}{name of the line}
}
}
\source{
generated using R
}
\usage{
lines
}
\description{
Lines
}
\keyword{datasets}
