% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mower_sensor}
\alias{mower_sensor}
\title{Mower sensor data in csv}
\format{
## `mower_sensor`
A data frame with 420 rows and 3 columns:
\describe{
  \item{LNG}{Longitude}
  \item{LAT}{Latitude}
  \item{ELEV}{Elevation}
  ...
}
}
\usage{
mower_sensor
}
\description{
Mower sensor data in csv
}
\keyword{datasets}
