% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NDRE}
\alias{NDRE}
\title{NDRE values}
\format{
stars
A stars object with 2 dimensions and 1 attribute:
}
\source{
privately obtained
}
\usage{
NDRE
}
\description{
NDRE values
}
\keyword{datasets}
