% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ne_counties}
\alias{ne_counties}
\title{Nebraska county boundary data}
\format{
`sf`
An sf with 2 rows and 1 columns:
\describe{
  \item{statefp}{State FIP}
  \item{countyfp}{County FIP}
  \item{name}{County name}
  \item{geometry}{geometry}
}
}
\source{
<https://www.nebraskamap.gov/datasets/nebraska::county-boundaries-census/about>
}
\usage{
ne_counties
}
\description{
Nebraska county boundary as an sf object
}
\keyword{datasets}
