% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NIR}
\alias{NIR}
\title{Near-Infrared Red (NIR)}
\format{
stars
A stars object with 2 dimensions and 1 attribute:
\describe{
  \item{NIR}{NIR}
}
}
\source{
privately obtained
}
\usage{
NIR
}
\description{
Near-Infrared Red from drone
}
\keyword{datasets}
