% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points}
\alias{points}
\title{Points}
\format{
sf
An sf with 3 rows and 2 columns:
\describe{
  \item{x}{geometry}
  \item{point_name}{name of the point}
}
}
\source{
generated using R
}
\usage{
points
}
\description{
Points
}
\keyword{datasets}
