% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygons}
\alias{polygons}
\title{Polygons}
\format{
sf
A data frame with 3 rows and 2 columns:
\describe{
  \item{x}{geometry}
  \item{polygon_name}{name of the polygon}
}
}
\source{
generated using R
}
\usage{
polygons
}
\description{
Polygons
}
\keyword{datasets}
