% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prism_2012_aug}
\alias{prism_2012_aug}
\title{PRISM precipitation data}
\format{
stars
A stars object with 3 dimensions and 1 attribute:
}
\source{
privately obtained
}
\usage{
prism_2012_aug
}
\description{
PRISM precipitation data from Aug 1 to Aug 5 in 2012
}
\keyword{datasets}
