% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prism_douglas}
\alias{prism_douglas}
\title{PRISM maximum temperature data for the Douglas county in NE}
\format{
stars
A stars object with 2 dimensions and 1 attribute:
}
\source{
privately obtained
}
\usage{
prism_douglas
}
\description{
PRISM maximum temperature data for the Douglas county in NE on 08/01/2012
}
\keyword{datasets}
