% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prism_us}
\alias{prism_us}
\title{PRISM precipitation data for the contiguous US}
\format{
stars
A stars object with 2 dimensions and 1 attribute:
}
\source{
privately obtained
}
\usage{
prism_us
}
\description{
PRISM precipitation data for the contiguous US on 08/01/2012
}
\keyword{datasets}
