% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{railroads_ne}
\alias{railroads_ne}
\title{US railroads in Nebraska}
\format{
`sf`
An sf with 2 rows and 1 columns:
\describe{
  \item{LINEARID}{rail ID}
  \item{geometry}{geometry}
}
}
\source{
<"http://www2.census.gov/geo/tiger/TIGER2019/RAILS/tl_2019_us_rails.zip">
}
\usage{
railroads_ne
}
\description{
US railroads stored as an sf object
}
\keyword{datasets}
