% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RED}
\alias{RED}
\title{Red (RED)}
\format{
stars
A stars object with 2 dimensions and 1 attribute:
}
\source{
privately obtained
}
\usage{
RED
}
\description{
Red from drone
}
\keyword{datasets}
