% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reflec_green}
\alias{reflec_green}
\title{Green reflectance}
\format{
stars
A stars object with 2 dimensions and 1 attribute:
\describe{
  \item{green}{green reflectance}
}
}
\source{
generated using R
}
\usage{
reflec_green
}
\description{
Green reflectance
}
\keyword{datasets}
