% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reflec_red}
\alias{reflec_red}
\title{Red reflectance}
\format{
stars
A stars object with 2 dimensions and 1 attribute:
\describe{
  \item{red}{red reflectance}
}
}
\source{
generated using R
}
\usage{
reflec_red
}
\description{
Red reflectance
}
\keyword{datasets}
