% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment_blocks}
\alias{treatment_blocks}
\title{Treatment blocks}
\format{
sf
An sf object with 25 rows and 4 columns:
\describe{
  \item{product}{product name}
  \item{target rate}{target nitrogen rate}
  \item{replication}{replication id}
  \item{geometry}{geometry}
}
}
\source{
privately obtained
}
\usage{
treatment_blocks
}
\description{
Treatment blocks
}
\keyword{datasets}
