% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wells_ne}
\alias{wells_ne}
\title{Wells in Nebraska with geographic coordinates}
\format{
## `wells_ne`
A data frame with 1,000 rows and 3 columns:
\describe{
  \item{wellid}{Well ID number}
  \item{longdd}{Longitude}
  \item{latdd}{Latitude}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
wells_ne
}
\description{
A subset of wells registered in Nebraska
}
\keyword{datasets}
