% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wells_ne_sf}
\alias{wells_ne_sf}
\title{Wells in Nebraska as sf}
\format{
sf
An sf with 1,000 rows and 2 columns:
\describe{
  \item{wellid}{Well ID number}
  \item{geometry}{geometry}
}
}
\source{
<https://dnr.nebraska.gov/groundwater>
}
\usage{
wells_ne_sf
}
\description{
This is an sf version of wells_ne
}
\keyword{datasets}
