% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{Country boundary}
\format{
sf
A data frame with 242 rows and 2 columns:
\describe{
  \item{name}{Country name}
  \item{geometry}{geometry}
}
}
\source{
rnaturalearhdata package
}
\usage{
world
}
\description{
Counties boundary as sf
}
\keyword{datasets}
