% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reflec_blue}
\alias{reflec_blue}
\title{Blue reflectance}
\format{
stars
A stars object with 2 dimensions and 1 attribute:
\describe{
  \item{blue}{blue reflectance}
}
}
\source{
generated using R
}
\usage{
reflec_blue
}
\description{
Blue reflectance
}
\keyword{datasets}
