% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soy_yield}
\alias{soy_yield}
\title{Soybean yield}
\format{
sf
A data frame with 1,683 rows and 3 columns:
\describe{
  \item{yield}{soybean yield (bu/acre)}
  \item{geometry}{geometry}
   \item{yield_id}{ID}
}
}
\source{
privately obtained
}
\usage{
soy_yield
}
\description{
Soybean yield (bu/acre)
}
\keyword{datasets}
